修正求职/职位相关列表奖金金额

+++ src/main/java/com/korant/youya/workplace/pojo/vo/huntjob/HuntJobDetailsPreviewVo.java
@@ -4,6 +4,7 @@ import com.korant.youya.workplace.annotations.Dict;
 import com.korant.youya.workplace.pojo.vo.workexperience.WorkExperiencePreviewVo;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -75,7 +76,7 @@ public class HuntJobDetailsPreviewVo {
     /**
      * 入职成功奖金
      */
-    private Integer onboardingAward;
+    private BigDecimal onboardingAward;
 
     /**
      * 求职状态
